pyproject.toml
[project]
name = "ai-middleware"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.0",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "mistralai>=1.9.11",
    "neo4j>=5.28.2",
    "neo4j-graphrag>=1.10.1",
    "pydantic-settings>=2.11.0",
    "ruff>=0.14.4",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
]
